Refactor names to closer match dispatching
