+++ config/dev.exs
@@ -27,7 +27,7 @@ config :faros, Faros.Endpoint,
 # Do not include metadata nor timestamps in development logs
 config :logger, :console, format: "[$level] $message\n"
 
-config :faros, :book_finder, Faros.Books.Finder
+config :faros, :book_finder, Faros.Books.GoogleFinder
 
 config :faros, Faros.Repo,
   adapter: Ecto.Adapters.MySQL,
