--- lib/faros/books/search_by_isbn.ex
@@ -1,6 +1,0 @@
-defmodule Faros.Books.SearchByIsbn do
-
-  def finder do
-    Application.get_env(:faros, :book_finder)
-  end
-end
